[FE] REFACTOR: Seperate MapSelect Component #747
