+++ frontend_v4/src/containers/MapInfoContainer.tsx
@@ -1,97 +1,11 @@
-import React, { useState } from "react";
+import { useState } from "react";
 import styled from "styled-components";
 import exitButton from "@/assets/images/exitButton.svg";
 import MapGridContainer from "./MapGridContainer";
 import useDetailInfo from "@/hooks/useDetailInfo";
 import { useRecoilValue } from "recoil";
 import { currentLocationFloorState } from "@/recoil/selectors";
-
-const SelectContainer = ({
-  floor,
-  setFloor,
-  floorInfo,
-}: {
-  floor: number;
-  setFloor: React.Dispatch<React.SetStateAction<number>>;
-  floorInfo: number[];
-}) => {
-  const onClickFloorOption = () => {
-    document.getElementById("mapFloorOptionBox")?.classList.toggle("on");
-  };
-
-  const selectFloor = (info: string) => {
-    const floorInfo = parseInt(info);
-    onClickFloorOption();
-    setFloor(floorInfo);
-  };
-
-  return (
-    <div style={{ position: "relative" }}>
-      <CurrentFloorStyled onClick={onClickFloorOption}>
-        {`${floor}층`}
-      </CurrentFloorStyled>
-      <OptionsContainer selectFloor={selectFloor} floorInfo={floorInfo} />
-    </div>
-  );
-};
-
-const CurrentFloorStyled = styled.div`
-  background: url("src/assets/images/select.svg") var(--main-color) no-repeat
-    80% 55%;
-  color: white;
-  cursor: pointer;
-  width: 65px;
-  height: 35px;
-  line-height: 35px;
-  text-indent: 12px;
-  border-radius: 10px;
-  margin-bottom: 50px;
-  &:hover {
-    opacity: 0.9;
-  }
-`;
-
-const OptionsContainerStyled = styled.div`
-  position: absolute;
-  left: 0;
-  top: 40px;
-  background: white;
-  border-radius: 10px;
-  box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.2);
-  overflow: hidden;
-  z-index: 99;
-  display: none;
-  &.on {
-    display: block;
-  }
-`;
-
-const OptionStyled = styled.div`
-  width: 65px;
-  height: 35px;
-  border-bottom: 1px solid #e6e6e6;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: var(--white);
-  color: black;
-  cursor: pointer;
-`;
-
-const OptionsContainer: React.FC<{
-  selectFloor: Function;
-  floorInfo: number[];
-}> = ({ floorInfo, selectFloor }) => {
-  return (
-    <OptionsContainerStyled id="mapFloorOptionBox">
-      {floorInfo.map((info, idx) => (
-        <OptionStyled onClick={() => selectFloor(info)} key={idx}>
-          {info}층
-        </OptionStyled>
-      ))}
-    </OptionsContainerStyled>
-  );
-};
+import MapFloorSelect from "@/components/MapSelect";
 
 const MapInfoContainer = () => {
   const { closeMap } = useDetailInfo();
@@ -107,11 +21,7 @@ const MapInfoContainer = () => {
           style={{ width: "24px", cursor: "pointer" }}
         />
       </HeaderStyled>
-      <SelectContainer
-        floor={floor}
-        setFloor={setFloor}
-        floorInfo={floorInfo}
-      />
+      <MapFloorSelect floor={floor} setFloor={setFloor} floorInfo={floorInfo} />
       <MapGridContainer floor={floor} />
     </MapInfoContainerStyled>
   );
@@ -119,18 +29,6 @@ const MapInfoContainer = () => {
 
 const H2Styled = styled.h2`
   font-size: 1.5rem;
-`;
-
-const SelectStyled = styled.select`
-  background: var(--main-color);
-  padding: 5px 10px;
-  border-radius: 5px;
-  margin-bottom: 50px;
-  & > option {
-    background: var(--main-color);
-    padding: 5px 10px;
-    border-radius: 5px;
-  }
 `;
 
 const HeaderStyled = styled.div`
